requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
openpyxl==3.1.2
python-dotenv==1.0.0
langchain==1.0.2
langchain-community==0.3.31
langchain-openai==1.0.1
langchain-chroma==0.2.6
langchain-huggingface==1.0.0
chromadb==1.1.1
sentence-transformers==2.3.1
pypdf==4.0.1
geopy==2.4.1
gspread==5.12.3
oauth2client==4.1.3